fix(daemon): record the interactive console session instead of session 0 on Windows
